change dialog to snackbar in ReadPlaylistFragment
